lib/posix-fd: Move open file headers into own lib
This changes merges the headers uk/ofile.h and uk/posix-fd.h and moves
them into their own library, providing shared definitions related to
open file descriptions for both posix-fdio and posix-fdtab, without
introducing an undue dedependency between the two.

Checkpatch-Ignore: REPEATED_WORD
