+++ lib/Makefile.uk
@@ -11,6 +11,7 @@ $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/ukgcov))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/isrlib))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/nolibc))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/posix-environ))
+$(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/posix-fd))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/posix-fdtab))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/posix-fdio))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/posix-eventfd))
