+++ lib/posix-fd/Makefile.uk
@@ -1,0 +1,4 @@
+$(eval $(call addlib_s,libposix_fd,$(CONFIG_LIBPOSIX_FD)))
+
+CINCLUDES-$(CONFIG_LIBPOSIX_FD) += -I$(LIBPOSIX_FD_BASE)/include
+CXXINCLUDES-$(CONFIG_LIBPOSIX_FD) += -I$(LIBPOSIX_FD_BASE)/include
