+++ lib/posix-fdio/Config.uk
@@ -2,6 +2,7 @@ config LIBPOSIX_FDIO
 	bool "posix-fdio: File I/O and control"
 	select LIBUKATOMIC
 	select LIBUKFILE
+	select LIBPOSIX_FD
 	select LIBPOSIX_TIME
 	select LIBUKTIMECONV
 	select LIBPOSIX_FDTAB
