+++ lib/ukfile/Config.uk
@@ -2,7 +2,6 @@ config LIBUKFILE
 	bool "ukfile: Common support for files"
 	select LIBUKATOMIC
 	select LIBUKLOCK
-	select LIBUKLOCK_MUTEX
 	select LIBUKLOCK_RWLOCK
 	select LIBUKSCHED
 	select LIBNOLIBC if !HAVE_LIBC
@@ -13,3 +12,4 @@ config LIBUKFILE_CHAINUPDATE
 
 config LIBUKFILE_FINALIZERS
 	bool
+	select LIBUKLOCK_MUTEX
